Remove unused modules and refactor playlist processing logic
Eliminated redundant scripts, consolidated functionality, and replaced scattered implementations with `PlaylistSplitter`. This cleanup improves maintainability by reducing duplication and enhancing clarity.
