+++ app/ProcessEnum.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class ProcessEnum(Enum):
+    WAITING = "waiting"
+    PROCESSING = "processing"
+    BREAK = "break"
+    COMPLETED = "completed"
+    ERROR = "error"
